Do layout to more pretty and add some moca data

+++ app/src/main/java/com/yuyu/barhopping/repository/datasource/FirebaseDataSource.kt
@@ -120,7 +120,7 @@ class FirebaseDataSource(context: Context) : FirebaseRepository {
                         endLon = document["endLon"] as String,
                         marketCount = document["marketCount"] as Long,
                         length = document["length"] as Long,
-                        hardDegree = document["hardDegree"] as Long?,
+                        hardDegree = document["hardDegree"] as Int?,
                         comments = document["comments"] as String?,
                         points = document["points"] as List<String>,
                         paths = document["paths"] as List<String>
